publishing events to all stream and renaming

+++ src/EventStore.Core/Commands/Transactions/ExponentialUnitOfWorkRetryOptions.cs
@@ -1,3 +1,3 @@
 ﻿namespace EventStore.Commands.Transactions;
 
-public record ExponentialUnitOfWorkRetryOptions(TimeSpan RetryInterval, int MaxRetries, double Exponential) : UnitOfWorkRetryOptions(RetryInterval, MaxRetries);
+public record ExponentialUnitOfWorkRetryOptions(TimeSpan RetryInterval, int MaxRetries, int Exponential) : UnitOfWorkRetryOptions(RetryInterval, MaxRetries);
